+++ src/components/molecules/ChartRangeSelector/ChartRangeSelector.tsx
@@ -20,24 +20,16 @@ import { useForm } from "react-hook-form";
 import { useTranslation } from "@neomanis/neo-translation";
 import { IconChevron } from "@/img/svg";
 import { getDateFnsLocaleFromUserLang } from "@/utils/dateTools";
-import NeoColors from "@/utils/neoColors";
 import { InputDateTime } from "@/components/atoms";
+import { RangeDateValue } from "@neomanis/neo-types";
+import { classNames } from "@/utils";
 
 export interface Props {
-    fCallBackData: (dates: { period: string | undefined; dates: { start: Date; end: Date } }) => void;
+    fCallBackData: (dates: { period: RangeDateValue; dates: { start: Date; end: Date } }) => void;
     fullSelector?: boolean;
     containerClassName?: string;
-    defaultValue?: { period: string | undefined; dates?: { start: Date; end: Date } };
+    defaultValue?: { period: RangeDateValue | undefined; dates?: { start: Date; end: Date } };
     resetDates?: { reset: boolean; setter: (val: boolean) => void };
-}
-
-enum rangeDateValue {
-    daily = "daily",
-    weekly = "weekly",
-    monthly = "monthly",
-    quarterly = "quarterly",
-    yearly = "yearly",
-    custom = "custom",
 }
 
 const ChartRangeSelector = ({
@@ -48,22 +40,20 @@ const ChartRangeSelector = ({
     resetDates,
 }: Props): ReactElement => {
     const { t, i18n } = useTranslation();
-    const data = [
-        { label: t("date.shortDateSelector.week"), value: rangeDateValue.weekly },
-        { label: t("date.shortDateSelector.month"), value: rangeDateValue.monthly },
-        { label: t("date.shortDateSelector.quarter"), value: rangeDateValue.quarterly },
-        { label: t("date.shortDateSelector.year"), value: rangeDateValue.yearly },
+    const data: { label: string; value: RangeDateValue }[] = [
+        { label: t("date.shortDateSelector.week"), value: "weekly" },
+        { label: t("date.shortDateSelector.month"), value: "monthly" },
+        { label: t("date.shortDateSelector.quarter"), value: "quarterly" },
+        { label: t("date.shortDateSelector.year"), value: "yearly" },
     ];
 
     const refDate: Date = useMemo(() => new Date(), []);
     if (fullSelector) {
-        data.unshift({ label: t("date.day_one"), value: rangeDateValue.daily });
-        data.push({ label: t("global.period"), value: rangeDateValue.custom });
+        data.unshift({ label: t("date.day_one"), value: "daily" });
+        data.push({ label: t("global.period"), value: "custom" });
     }
 
-    const [typeRangeSelect, setTypeRangeSelect] = useState<rangeDateValue | undefined>(
-        rangeDateValue[defaultValue.period as rangeDateValue]
-    );
+    const [typeRangeSelect, setTypeRangeSelect] = useState<RangeDateValue>(defaultValue.period);
     const [offset, setOffset] = useState<number>(0);
 
     const [customRange, setCustomRange] = useState<{ start: Date; end: Date }>({
@@ -130,17 +120,17 @@ const ChartRangeSelector = ({
 
     const dateRange = useMemo(() => {
         switch (typeRangeSelect) {
-            case rangeDateValue.daily:
+            case "daily":
                 return dayRangePicker(defaultValue.dates?.start ?? refDate, offset);
-            case rangeDateValue.weekly:
+            case "weekly":
                 return weekRangePicker(defaultValue.dates?.start ?? refDate, offset);
-            case rangeDateValue.monthly:
+            case "monthly":
                 return monthRangePicker(defaultValue.dates?.start ?? refDate, offset);
-            case rangeDateValue.quarterly:
+            case "quarterly":
                 return quarterRangePicker(defaultValue.dates?.start ?? refDate, offset);
-            case rangeDateValue.yearly:
+            case "yearly":
                 return yearRangePicker(defaultValue.dates?.start ?? refDate, offset);
-            case rangeDateValue.custom:
+            case "custom":
                 return {
                     start: isDate(customRange.start) ? customRange.start : startOfDay(refDate),
                     end: isDate(customRange.end) ? customRange.end : endOfDay(refDate),
@@ -152,19 +142,19 @@ const ChartRangeSelector = ({
 
     const textShow = useMemo(() => {
         switch (typeRangeSelect) {
-            case rangeDateValue.daily:
+            case "daily":
                 return format(dateRange.start, "EEEE P", {
                     locale: getDateFnsLocaleFromUserLang(i18n.language),
                 });
-            case rangeDateValue.weekly:
+            case "weekly":
                 return `${t("date.shortDateSelector.week")} ${format(dateRange.start, "I yyyy")}`;
-            case rangeDateValue.monthly:
+            case "monthly":
                 return format(dateRange.start, "MMMM yyyy", {
                     locale: getDateFnsLocaleFromUserLang(i18n.language),
                 });
-            case rangeDateValue.quarterly:
+            case "quarterly":
                 return t("date.shortDateSelector.quarter").charAt(0) + format(dateRange.start, "Q yyyy");
-            case rangeDateValue.yearly:
+            case "yearly":
                 return dateRange.start.getFullYear().toString();
             default:
                 return undefined;
@@ -189,22 +179,22 @@ const ChartRangeSelector = ({
                 start: refDate,
                 end: refDate,
             });
-            setTypeRangeSelect(undefined);
+            setTypeRangeSelect(defaultValue.period);
             resetDates.setter(false);
         }
     }, [resetDates]);
 
     function getWidth(): number {
         switch (typeRangeSelect) {
-            case rangeDateValue.daily:
+            case "daily":
                 return 170;
-            case rangeDateValue.weekly:
+            case "weekly":
                 return 130;
-            case rangeDateValue.monthly:
+            case "monthly":
                 return 120;
-            case rangeDateValue.quarterly:
+            case "quarterly":
                 return 70;
-            case rangeDateValue.yearly:
+            case "yearly":
                 return 50;
             default:
                 return 100;
@@ -234,14 +224,14 @@ const ChartRangeSelector = ({
                 className="relative flex text-neo-blue-secondary font-bold text-sm"
             >
                 {data.map((item, key) => (
-                    <li className="mx-2 flex items-center" key={key}>
+                    <li className="mx-2 flex items-center transition-all" key={key}>
                         <p
-                            className={`uppercase transition-colors rounded-full px-4 py-1
-                        ${
-                            item.value === typeRangeSelect
-                                ? "bg-neo-blue text-white"
-                                : "cursor-pointer hover:bg-neo-blue-secondary hover:text-white"
-                        }`}
+                            className={classNames(
+                                "uppercase rounded-full px-4 py-1 transition-all",
+                                item.value === typeRangeSelect
+                                    ? "bg-neo-blue text-white"
+                                    : "cursor-pointer hover:bg-neo-blue-secondary hover:text-white hover:scale-110"
+                            )}
                             onClick={() => {
                                 setTypeRangeSelect(item.value);
                                 setOffset(0);
@@ -249,9 +239,9 @@ const ChartRangeSelector = ({
                         >
                             {item.label}
                         </p>
-                        {item.value === typeRangeSelect && item.value !== rangeDateValue.custom && (
+                        {item.value === typeRangeSelect && item.value !== "custom" && (
                             <div className="flex items-center">
-                                <p className={`mx-2 text-white capitalize`} style={{ width: getWidth() }}>
+                                <p className="mx-2 text-white capitalize" style={{ width: getWidth() }}>
                                     {textShow}
                                 </p>
                                 <div>
@@ -259,25 +249,37 @@ const ChartRangeSelector = ({
                                         onClick={() => {
                                             showUpDate() && setOffset(offset + 1);
                                         }}
-                                        className={`transform rotate-180 mb-2 p-1
-                                    ${showUpDate() ? "cursor-pointer hover:scale-110 transition-all " : "opacity-30"}`}
+                                        className={classNames(
+                                            "transform rotate-180 mb-2 p-1 group",
+                                            showUpDate()
+                                                ? "cursor-pointer hover:scale-110 transition-all"
+                                                : "opacity-30"
+                                        )}
                                     >
                                         {/* caret up */}
-                                        <IconChevron width={15} fill={NeoColors.blue.secondary} />
+                                        <IconChevron
+                                            className={classNames(
+                                                showUpDate() &&
+                                                    "opacity-60 group-hover:opacity-100 group-hover:animate-bounce",
+                                                "w-4 fill-neo-blue-secondary"
+                                            )}
+                                        />
                                     </div>
                                     <div
                                         onClick={() => {
                                             setOffset(offset - 1);
                                         }}
-                                        className="transform hover:scale-110 transition-all hover:cursor-pointer mt-2 p-1"
+                                        className="transform hover:scale-110 rotate-180 transition-all hover:cursor-pointer mt-2 p-1 group "
                                     >
                                         {/* caret down  */}
-                                        <IconChevron width={15} fill={NeoColors.blue.secondary} />
+                                        <div className="rotate-180">
+                                            <IconChevron className="w-4 fill-neo-blue-secondary group-hover:animate-bounce opacity-60 group-hover:opacity-100 transition-opacity" />
+                                        </div>
                                     </div>
                                 </div>
                             </div>
                         )}
-                        {item.value === typeRangeSelect && item.value === rangeDateValue.custom && (
+                        {item.value === typeRangeSelect && item.value === "custom" && (
                             <InputDateTime
                                 className="w-72 px-4 z-50"
                                 defaultValue={periodDefaultValue}