+++ changelog.md
@@ -1,6 +1,16 @@
 # Changelog 
 
- ### [5.22.0] 
+ ### [5.26.0] 
+ add annim in grid and add hover in input component and button icon  ### [5.25.4] 
+ :rotating_light: fixed some memory leak and key react warning ### [5.25.3] 
+ :adhesive_bandage: fixed time spent input width and :poop: typescript, i know where you hide, come here so i can kill you motherfucker ### [5.25.2] 
+ :recycle: updated chart range selector to match time spent screen, fixed issue with useOnClickOutside data attribute exception ### [5.25.1] 
+ :adhesive_bandage: fixed closing handle of time spent input and button add/remove visual ### [5.25.0] 
+ :sparkles: added time spent component ### [5.24.2] 
+ fix updater problem on display or hide ### [5.24.1] 
+ handle correctly index of glpi ticket priority and scale ### [5.24.0] 
+ :sparkles: added a sortable table component ### [5.23.0] 
+ add className props on ul element in shadowBoxWrapper to allow custom spacing ### [5.22.0] 
  delete some priority type and update color and function of priority ### [5.21.2] 
  change close color value and get status color not priority for new ticket ### [5.21.1] 
  repair test text for uneditableField ### [5.21.0]